add base controller api, and appsetting con string scope
