+++ Src/CleanArchitectureProjectTemplate.WebApi/Controllers/v1/WeatherForecastController.cs
@@ -1,0 +1,32 @@
+using Asp.Versioning;
+using Microsoft.AspNetCore.Mvc;
+
+namespace CleanArchitectureProjectTemplate.WebApi.Controllers.v1;
+
+[ApiVersion(1.0)]
+public class WeatherForecastController : BaseApiController
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    private readonly ILogger<WeatherForecastController> _logger;
+
+    public WeatherForecastController(ILogger<WeatherForecastController> logger)
+    {
+        _logger = logger;
+    }
+
+    [HttpGet(Name = "GetWeatherForecast")]
+    public IEnumerable<WeatherForecast> Get()
+    {
+        return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+        {
+            Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+            TemperatureC = Random.Shared.Next(-20, 55),
+            Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+        })
+        .ToArray();
+    }
+}
