+++ Src/CleanArchitectureProjectTemplate.WebApi/appsettings.json
@@ -1,4 +1,8 @@
 {
+  "ConnectionStrings": {
+    "Development": "",
+    "Production": ""
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
